final styles
